removing GCP presence. renaming package to mirror name change:
