+++ build.gradle
@@ -18,7 +18,7 @@ apply plugin: 'io.spring.dependency-management'
 apply plugin: 'checkstyle'
 apply plugin: 'jacoco'
 
-group 'com.broadcastapi'
+group 'com.earlybirdapi'
 version '1.0-SNAPSHOT'
 
 sourceCompatibility = 1.8
@@ -40,11 +40,8 @@ dependencies {
     compile(group: 'io.springfox', name: 'springfox-swagger-ui', version: '2.9.2')
 
     compile(group: 'org.springframework.boot', name: 'spring-boot-starter-data-jpa', version: '2.0.5.RELEASE')
-    compile(group: 'mysql', name: 'mysql-connector-java', version: '8.0.11')
-    compile(group: 'com.google.cloud.sql', name: 'mysql-socket-factory-connector-j-8', version: '1.0.11')
     compile(group: 'com.h2database', name: 'h2', version: '1.4.197')
 
-    compile group: 'com.google.code.gson', name: 'gson', version: '2.8.5'
     compile group: 'org.springframework.security', name: 'spring-security-core', version: '5.1.2.RELEASE'
 
     compile(group: 'org.projectlombok', name: 'lombok', version: '1.18.2') {
@@ -70,9 +67,9 @@ jacocoTestReport {
     afterEvaluate {
         classDirectories = files(classDirectories.files.collect {
             fileTree(dir: it,
-                    exclude: ['com/lhd/broadcastapi/util/**',
-                              'com/lhd/broadcastapi/config/**',
-                              'com/lhd/broadcastapi/BroadcastApplication.class'])
+                    exclude: ['com/lhd/earlybirdapi/util/**',
+                              'com/lhd/earlybirdapi/config/**',
+                              'com/lhd/earlybirdapi/BroadcastApplication.class'])
         })
     }
 }
@@ -91,9 +88,9 @@ jacocoTestCoverageVerification {
     afterEvaluate {
         classDirectories = files(classDirectories.files.collect {
             fileTree(dir: it,
-                    exclude: ['com/lhd/broadcastapi/util/**',
-                              'com/lhd/broadcastapi/config/**',
-                              'com/lhd/broadcastapi/BroadcastApplication.class'])
+                    exclude: ['com/lhd/earlybirdapi/util/**',
+                              'com/lhd/earlybirdapi/config/**',
+                              'com/lhd/earlybirdapi/BroadcastApplication.class'])
         })
     }
 }
