+++ settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'broadcastapi'
+rootProject.name = 'earlybirdapi'
 include 'src'
