+++ src/main/java/com/lhd/earlybirdapi/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.lhd.broadcastapi.config;
+package com.lhd.earlybirdapi.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
