+++ src/main/java/com/lhd/earlybirdapi/subscription/GithubRepoRepository.java
@@ -1,4 +1,4 @@
-package com.lhd.broadcastapi.subscription;
+package com.lhd.earlybirdapi.subscription;
 
 import org.springframework.data.jpa.repository.JpaRepository;
