+++ src/main/java/com/lhd/earlybirdapi/subscription/IssueDto.java
@@ -1,8 +1,7 @@
-package com.lhd.broadcastapi.subscription;
+package com.lhd.earlybirdapi.subscription;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import java.time.Instant;
 import lombok.Data;
 import lombok.NoArgsConstructor;
