--- src/main/webapp/WEB-INF/appengine-web.xml
@@ -1,6 +1,0 @@
-<appengine-web-app xmlns="http://appengine.google.com/ns/1.0">
-  <version>1</version>
-  <threadsafe>true</threadsafe>
-  <runtime>java8</runtime>
-  <service>broadcast-api</service>
-</appengine-web-app>
